+++ semcore/fullscreen-modal/__tests__/fullscreen-modal.axe-test.tsx
@@ -1,13 +1,10 @@
-import { e2eStandToHtml } from '@semcore/testing-utils/e2e-stand';
 import { expect, getAccessibilityViolations, test } from '@semcore/testing-utils/playwright';
+import { loadPage } from '@semcore/testing-utils/shared/helpers';
+import { TAG } from '@semcore/testing-utils/shared/tags';
 
-test.describe('Fullscreen modal', () => {
+test.describe(`${TAG.ACCESSIBILITY} @fullscreen-modal `, () => {
   test('Basic usage', async ({ page }) => {
-    const standPath =
-      'stories/components/fullscreen-modal/docs/examples/basic_fullscreen.headers_use.tsx';
-    const htmlContent = await e2eStandToHtml(standPath, 'en');
-
-    await page.setContent(htmlContent);
+    await loadPage(page, 'stories/components/fullscreen-modal/docs/examples/basic_fullscreen.headers_use.tsx', 'en');
 
     await page.locator('[data-ui-name="Button"]').click();
     await page.waitForTimeout(200);
@@ -17,11 +14,7 @@ test.describe('Fullscreen modal', () => {
   });
 
   test('Dual zone', async ({ page }) => {
-    const standPath =
-      'stories/components/fullscreen-modal/docs/examples/example_of_a_dual-zone_modal_window.tsx';
-    const htmlContent = await e2eStandToHtml(standPath, 'en');
-
-    await page.setContent(htmlContent);
+    await loadPage(page, 'stories/components/fullscreen-modal/docs/examples/example_of_a_dual-zone_modal_window.tsx', 'en');
 
     await page.locator('[data-ui-name="Button"]').click();
     await page.waitForTimeout(200);