+++ semcore/data-table/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 ### Fixed
 
 - The focus did not move to the trigger via `Escape` when opening the accordion with the mouse.
+- SR didn't pronounce all items selecting/deselecting on table with pagination.
 
 ## [16.4.3] - 2025-10-31